# dist/config/swagger.json
{
    "swagger": "2.0",
    "title": "Invoice Builder Api Documentation",
    "description": "Swagger API Documentation for Invoice Builder MEAN Stack App",
    "license": {
        "name": "Apache 2.0",
        "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
    },
    "version": "2.0",
    "host" : "localhost:3000",
    "basePath" : "/api",
    "tags" : [
        {
            "name" : "Invoices",
            "description" : "API for Invoices"
        }
    ],
    "schemes" : ["http"],
    "consumes" : ["application/json"],
    "produces" : ["application/json"],
    "paths" : {
        "/invoices" : {
            "post" : {
                "tags" : ["Invoices"],
                "description" : "Create a new Invoice",
                "summary" : "Create a new Invoice",
                "parameters" : [
                    {
                        "name" : "invoice",
                        "description" : "invoice about to be created",
                        "in" : "body",
                        "required" : true,
                        "schema" : {
                            "$ref" : "#definitions/Invoice"
                        }
                    }
                ],
                "produces" : ["application/json"],
                "responses" : {
                    "200" : {
                        "description" : "new invoice created",
                        "schema" : {
                            "$ref" : "#definitions/Invoice"
                        }
                    }
                }
            },
            "get" : {
                "tags" : ["Invoices"],
                "summary" : "Find all invoices from server",
                "description" : "All Invoices",
                "parameters" : [{
                    "name" : "page",
                    "type" : "integer",
                    "description" : "Set the page for pagination",
                    "in" : "query"
                },{
                    "name" : "perPage",
                    "type" : "integer",
                    "description" : "Define number of per page records",
                    "in" : "query"
                },{
                    "name" : "filter",
                    "type" : "string",
                    "description" : "Filter Invoices by item name",
                    "in" : "query"
                },{
                    "name" : "sortField",
                    "type" : "string",
                    "description" : "Define the sort field for sorting",
                    "in" : "query"
                },{
                    "name" : "sortDir",
                    "type" : "string",
                    "description" : "Define the sort sort direction for sorting in ascending/decending direction",
                    "in" : "query"
                }],
                "responses" : {
                    "200" : {
                        "description" : "OK",
                        "schema" : {
                            "$ref" : "#definitions/Invoices"
                        }
                    }
                }
            }
        },
        "/invoices/{id}" : {
            "parameters" : [
                {
                    "name" : "id",
                    "in" : "path",
                    "description" : "Enter invoice Id",
                    "type" :"string"
                }
            ],
            "get" : {
                "tags" : ["Invoices"],
                "description" : "Find Invoice by Id",
                "summary" : "Find single Invoice",
                "responses" : {
                    "200" : {
                        "description" : "Invoice Found",
                        "schema" : {
                            "$ref" : "#definitions/Invoice"
                        }
                    }
                  }
                },
            "delete" : {
                "tags" : ["Invoices"],
                "description" : "Delete Invoice by Id",
                "summary" : "Delete single Invoice",
                "responses" : {
                    "200" : {
                        "description" : "Invoice Deleted",
                        "schema" : {
                            "$ref" : "#definitions/Invoice"
                            }
                        }
                    }    
            },
            "put" : {
                "tags" : ["Invoices"],
                "description" : "Update an Invoice",
                "summary" : "Update an existing Invoice",
                "parameters" : [
                    {
                        "name" : "Invoice",
                        "description" : "invoice about to be updated",
                        "in" : "body",
                        "required" : true,
                        "schema" : {
                            "$ref" : "#definitions/UpdateInvoice"
                        }
                    }
                ],
                "produces" : ["application/json"],
                "responses" : {
                    "200" : {
                        "description" : "Invoice Updated",
                        "schema" : {
                            "$ref" : "#definitions/Invoice"
                        }
                    }
                }
            }
        }
        
    },
    "definitions" : {
        "Invoice" : {            
            "required" : ["item,qty,date,due"],
            "properties" : {
                "_id": {
                    "type": "string",
                    "uniqueItems" : true
                },
                "item": {
                    "type": "string"
                },
                "qty": {
                    "type": "integer",
                    "format" : "int32"
                },
                "date": {
                    "type": "string",
                    "format" : "date"
                },
                "due": {
                    "type": "string",
                    "format" : "date"
                },
                "rate": {
                    "type": "number",
                    "format" : "float"
                },
                "tax": {
                    "type": "number",
                    "format" : "float"
                }
            }
        },
        "UpdateInvoice" : {            
            "properties" : {
                "item": {
                    "type": "string"
                },
                "qty": {
                    "type": "integer",
                    "format" : "int32"
                },
                "date": {
                    "type": "string",
                    "format" : "date"
                },
                "due": {
                    "type": "string",
                    "format" : "date"
                },
                "rate": {
                    "type": "number",
                    "format" : "float"
                },
                "tax": {
                    "type": "number",
                    "format" : "float"
                }
            }
        }
    },
    "Invoices" : {
        "properties" : { 
            "docs" : {
              "type" : "array",
              "$ref" : "#definitions/Invoice"
              }
            },
        "total" : {
            "type" : "integer"
        },
        "page" : {
            "type" : "integer"
        },
        "pags" : {
            "type" : "integer"
        },
        "limit" : {
            "type" : "integer"
        }      
    }
  }